TST GH26807 split out test_extract
